Moving bot to db test to local hyperdrive test
