--- tests/bot_to_db_test.py
@@ -1,376 +1,0 @@
-"""System test for end to end usage of agent0 libraries."""
-
-from __future__ import annotations
-
-from decimal import Decimal
-from typing import Type
-
-import numpy as np
-import pandas as pd
-import pytest
-from fixedpointmath import FixedPoint, isclose
-
-from agent0.chainsync.db.hyperdrive.interface import get_pool_analysis, get_pool_config, get_pool_info
-from agent0.core.base.make_key import make_private_key
-from agent0.core.hyperdrive.interactive import LocalHyperdrive
-from agent0.core.test_utils import CycleTradesPolicy
-from agent0.ethpy.hyperdrive import BASE_TOKEN_SYMBOL
-
-
-def _to_unscaled_decimal(fp_val: FixedPoint) -> Decimal:
-    return Decimal(str(fp_val))
-
-
-class TestBotToDb:
-    """Test pipeline from bots making trades to viewing the trades in the db."""
-
-    # TODO split this up into different functions that work with tests
-    # pylint: disable=too-many-locals, too-many-statements
-    # ruff: noqa: PLR0915 (Too many statements)
-    @pytest.mark.anvil
-    def test_bot_to_db(
-        self,
-        fast_hyperdrive_fixture: LocalHyperdrive,
-        cycle_trade_policy: Type[CycleTradesPolicy],
-    ):
-        """Run the entire pipeline and checks the database at the end. All arguments are fixtures."""
-
-        # Initialize agent
-        private_key = make_private_key()
-        agent = fast_hyperdrive_fixture.init_agent(
-            private_key=private_key,
-            base=FixedPoint(1_000_000),
-            eth=FixedPoint(100),
-            policy=cycle_trade_policy,
-            policy_config=cycle_trade_policy.Config(
-                slippage_tolerance=FixedPoint("0.0001"),
-            ),
-        )
-
-        # Run trades
-        while not agent.policy_done_trading:
-            agent.execute_policy_action()
-
-        # Run bots again, but this time only for 3 trades
-        # TODO agent rework will allow us to separate policy from agent
-        # but for now, we need to reinitialize another agent with the same
-        # key to run the policy from scratch again
-        agent = fast_hyperdrive_fixture.init_agent(
-            private_key=private_key,
-            # We don't add funds here again, as the agent already has funds
-            policy=cycle_trade_policy,
-            policy_config=cycle_trade_policy.Config(
-                slippage_tolerance=FixedPoint("0.0001"),
-                max_trades=3,
-            ),
-        )
-
-        while not agent.policy_done_trading:
-            agent.execute_policy_action()
-
-        # This bot does the following known trades in sequence:
-        # 1. addLiquidity of 111_111 base
-        # 2. openLong of 22_222 base
-        # 3. openShort of 333 bonds
-        # 4. removeLiquidity of all LP tokens
-        # 5. addLiquidity of 111_111 base
-        # 6. closeLong on long from trade 2
-        # 7. closeShort on short from trade 3
-        # 8. redeemWithdrawalShares of all withdrawal tokens from trade 4
-        # The bot then runs again, this time for 3 trades:
-        # 9. addLiquidity of 111_111 base
-        # 10. openLong of 22_222 base
-        # 11. openShort of 333 bonds
-
-        # Test db entries are what we expect
-        # We don't coerce to float because we want exact values in decimal
-
-        db_pool_config_df: pd.DataFrame = get_pool_config(fast_hyperdrive_fixture.db_session, coerce_float=False)
-
-        # TODO these expected values are defined in src/agent0/ethpy/test_fixtures/deploy_hyperdrive.py
-        # Eventually, we want to parameterize these values to pass into deploying hyperdrive
-        initial_fixed_rate = FixedPoint("0.05")
-        # This expected time stretch is only true for 1 year position duration
-        expected_timestretch_fp = FixedPoint(1) / (
-            FixedPoint("5.24592") / (FixedPoint("0.04665") * (initial_fixed_rate * FixedPoint(100)))
-        )
-        expected_timestretch = _to_unscaled_decimal(expected_timestretch_fp)
-        expected_inv_timestretch = _to_unscaled_decimal((1 / expected_timestretch_fp))
-        # Ignore linker factory since we don't know the target address
-        db_pool_config_df = db_pool_config_df.drop(columns=["linker_factory"])
-        deployer_address = fast_hyperdrive_fixture.chain.get_deployer_account_address()
-        # pylint: disable=protected-access
-        base_token_addr = fast_hyperdrive_fixture._deployed_hyperdrive.base_token_contract.address
-        vault_shares_token_addr = fast_hyperdrive_fixture._deployed_hyperdrive.vault_shares_token_contract.address
-        expected_pool_config = {
-            "hyperdrive_address": fast_hyperdrive_fixture.get_hyperdrive_address(),
-            "base_token": base_token_addr,
-            "vault_shares_token": vault_shares_token_addr,
-            "initial_vault_share_price": _to_unscaled_decimal(FixedPoint("1")),
-            "minimum_share_reserves": _to_unscaled_decimal(FixedPoint("10")),
-            "minimum_transaction_amount": _to_unscaled_decimal(FixedPoint("0.001")),
-            "circuit_breaker_delta": _to_unscaled_decimal(FixedPoint("2")),
-            "position_duration": 60 * 60 * 24 * 365,  # 1 year
-            "checkpoint_duration": 3600,  # 1 hour
-            "time_stretch": expected_timestretch,
-            "governance": deployer_address,
-            "fee_collector": deployer_address,
-            "sweep_collector": deployer_address,
-            "curve_fee": _to_unscaled_decimal(FixedPoint("0.01")),  # 1%
-            "flat_fee": _to_unscaled_decimal(FixedPoint("0.0005")),  # 0.05% APR
-            "governance_lp_fee": _to_unscaled_decimal(FixedPoint("0.15")),  # 15%
-            "governance_zombie_fee": _to_unscaled_decimal(FixedPoint("0.03")),  # 3%
-            "inv_time_stretch": expected_inv_timestretch,
-        }
-
-        # Existence test
-        assert len(db_pool_config_df) == 1, "DB must have one entry for pool config"
-        db_pool_config: pd.Series = db_pool_config_df.iloc[0]
-
-        # Ensure keys match
-        # Converting to sets and compare
-        db_keys = set(db_pool_config.index)
-        expected_keys = set(expected_pool_config.keys())
-        assert db_keys == expected_keys, "Keys in db do not match expected"
-
-        # Value comparison
-        for key, expected_value in expected_pool_config.items():
-            assert_val = db_pool_config[key] == expected_value
-            assert assert_val, f"Values do not match for {key} ({db_pool_config[key]} != {expected_value})"
-
-        # Pool info comparison
-        db_pool_info: pd.DataFrame = get_pool_info(fast_hyperdrive_fixture.db_session, coerce_float=False)
-        expected_pool_info_keys = [
-            # Keys from contract call
-            "block_number",
-            "timestamp",
-            "share_reserves",
-            "share_adjustment",
-            "zombie_base_proceeds",
-            "zombie_share_reserves",
-            "bond_reserves",
-            "lp_total_supply",
-            "vault_share_price",
-            "longs_outstanding",
-            "long_average_maturity_time",
-            "shorts_outstanding",
-            "short_average_maturity_time",
-            "withdrawal_shares_ready_to_withdraw",
-            "withdrawal_shares_proceeds",
-            "lp_share_price",
-            "long_exposure",
-            # Added keys
-            "epoch_timestamp",
-            "total_supply_withdrawal_shares",
-            "gov_fees_accrued",
-            "hyperdrive_base_balance",
-            "hyperdrive_eth_balance",
-            "variable_rate",
-            "vault_shares",
-        ]
-        # Convert to sets and compare
-        assert set(db_pool_info.columns) == set(expected_pool_info_keys)
-
-        db_transaction_info: pd.DataFrame = get_transactions(fast_hyperdrive_fixture.db_session, coerce_float=False)
-        # TODO check transaction keys
-        # This likely involves cleaning up what columns we grab from transactions
-
-        db_wallet_delta: pd.DataFrame = get_wallet_deltas(fast_hyperdrive_fixture.db_session, coerce_float=False)
-
-        # Ensure trades exist in database
-        # Should be 11 total transactions
-        expected_number_of_transactions = 11
-        assert len(db_transaction_info) == expected_number_of_transactions
-        np.testing.assert_array_equal(
-            db_transaction_info["input_method"],
-            [
-                "addLiquidity",
-                "openLong",
-                "openShort",
-                "removeLiquidity",
-                "addLiquidity",
-                "closeLong",
-                "closeShort",
-                "redeemWithdrawalShares",
-                "addLiquidity",
-                "openLong",
-                "openShort",
-            ],
-        )
-
-        # 10 total trades in wallet deltas
-        assert db_wallet_delta["block_number"].nunique() == expected_number_of_transactions
-        # 23 different wallet deltas (2 token deltas per trade except for withdraw shares, which is 3)
-        assert len(db_wallet_delta) == 2 * expected_number_of_transactions + 1
-
-        actual_num_longs = Decimal("nan")
-        actual_num_shorts = Decimal("nan")
-        actual_num_lp = Decimal("nan")
-        actual_num_withdrawal = Decimal("nan")
-        # Go through each trade and ensure wallet deltas are correct
-        # The asserts here are equality because they are either int -> Decimal, which is lossless,
-        # or they're comparing values after the lossy conversion
-        expected_number_of_deltas = 2
-        for _, txn in db_transaction_info.iterrows():
-            # TODO differentiate between the first, second, and third addLiquidity
-            block_number = txn["block_number"]
-            if txn["input_method"] == "addLiquidity":
-                assert txn["input_params_contribution"] == Decimal(111_111)
-                # Filter for all deltas of this trade
-                block_wallet_deltas = db_wallet_delta[db_wallet_delta["block_number"] == block_number]
-                # Ensure number of token deltas
-                assert len(block_wallet_deltas) == expected_number_of_deltas
-                lp_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == "LP"]
-                base_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == BASE_TOKEN_SYMBOL]
-                assert len(lp_delta_df) == 1
-                assert len(base_delta_df) == 1
-                lp_delta = lp_delta_df.iloc[0]
-                base_delta = base_delta_df.iloc[0]
-                # 111_111 base for...
-                assert base_delta["delta"] == -Decimal(111_111)
-
-                # TODO check LP delta
-                # TODO check wallet info matches the deltas
-                # TODO check pool info after this tx
-
-                actual_num_lp = lp_delta["delta"]
-
-            # TODO differentiate between the first and second openLong
-            if txn["input_method"] == "openLong":
-                assert txn["input_params_amount"] == Decimal(22_222)
-                # Filter for all deltas of this trade
-                block_wallet_deltas = db_wallet_delta[db_wallet_delta["block_number"] == block_number]
-                # Ensure number of token deltas
-                assert len(block_wallet_deltas) == expected_number_of_deltas
-                long_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == "LONG"]
-                base_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == BASE_TOKEN_SYMBOL]
-                assert len(long_delta_df) == 1
-                assert len(base_delta_df) == 1
-                long_delta = long_delta_df.iloc[0]
-                base_delta = base_delta_df.iloc[0]
-                # 22_222 base for...
-                assert base_delta["delta"] == -Decimal(22_222)
-                # TODO check long delta
-                # TODO check maturity time and token_type
-                # TODO check current wallet matches the deltas
-                # TODO check pool info after this tx
-
-                actual_num_longs = long_delta["delta"]
-
-            # TODO differentiate between the first and second openShort
-            if txn["input_method"] == "openShort":
-                assert txn["input_params_bond_amount"] == Decimal(333)
-                block_wallet_deltas = db_wallet_delta[db_wallet_delta["block_number"] == block_number]
-                assert len(block_wallet_deltas) == expected_number_of_deltas
-                short_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == "SHORT"]
-                base_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == BASE_TOKEN_SYMBOL]
-                assert len(short_delta_df) == 1
-                assert len(base_delta_df) == 1
-                short_delta = short_delta_df.iloc[0]
-                base_delta = base_delta_df.iloc[0]
-                # 333 bonds for...
-                assert short_delta["delta"] == Decimal(333)
-                # TODO check base delta
-                # TODO check maturity time and token_type
-                # TODO check pool info after this tx
-
-                actual_num_shorts = short_delta["delta"]
-
-            if txn["input_method"] == "removeLiquidity":
-                # TODO change this to expected num lp
-                assert txn["input_params_lp_shares"] == actual_num_lp
-                block_wallet_deltas = db_wallet_delta[db_wallet_delta["block_number"] == block_number]
-                assert len(block_wallet_deltas) == expected_number_of_deltas + 1  # 3 deltas for withdraw shares
-                lp_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == "LP"]
-                withdrawal_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == "WITHDRAWAL_SHARE"]
-                base_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == BASE_TOKEN_SYMBOL]
-                assert len(lp_delta_df) == 1
-                assert len(withdrawal_delta_df) == 1
-                assert len(base_delta_df) == 1
-                lp_delta = lp_delta_df.iloc[0]
-                withdrawal_delta = withdrawal_delta_df.iloc[0]
-                base_delta = base_delta_df.iloc[0]
-                # TODO check against expected lp
-                assert lp_delta["delta"] == -actual_num_lp
-                # TODO check base delta
-                # TODO check withdrawal delta
-                # TODO check wallet info matches the deltas
-                # TODO check pool info after this tx
-
-                actual_num_withdrawal = withdrawal_delta["delta"]
-
-            if txn["input_method"] == "closeLong":
-                # TODO change this to expected long amount
-                assert txn["input_params_bond_amount"] == actual_num_longs
-                block_wallet_deltas = db_wallet_delta[db_wallet_delta["block_number"] == block_number]
-                assert len(block_wallet_deltas) == expected_number_of_deltas
-                long_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == "LONG"]
-                base_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == BASE_TOKEN_SYMBOL]
-                assert len(long_delta_df) == 1
-                assert len(base_delta_df) == 1
-                long_delta = long_delta_df.iloc[0]
-                base_delta = base_delta_df.iloc[0]
-                # TODO check against expected longs
-                assert long_delta["delta"] == -actual_num_longs
-                # TODO check base delta
-                # TODO check maturity time and token_type
-                # TODO check wallet info matches the deltas
-                # TODO check pool info after this tx
-
-            if txn["input_method"] == "closeShort":
-                assert txn["input_params_bond_amount"] == Decimal(333)
-                block_wallet_deltas = db_wallet_delta[db_wallet_delta["block_number"] == block_number]
-                assert len(block_wallet_deltas) == expected_number_of_deltas
-                short_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == "SHORT"]
-                base_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == BASE_TOKEN_SYMBOL]
-                assert len(short_delta_df) == 1
-                assert len(base_delta_df) == 1
-                short_delta = short_delta_df.iloc[0]
-                base_delta = base_delta_df.iloc[0]
-                # TODO check against expected shorts
-                assert short_delta["delta"] == -actual_num_shorts
-                # TODO check base delta
-                # TODO check maturity time and token_type
-                # TODO check wallet info matches the deltas
-                # TODO check pool info after this tx
-
-            if txn["input_method"] == "redeemWithdrawalShares":
-                # TODO change this to expected withdrawal shares
-                assert txn["input_params_withdrawal_shares"] == actual_num_withdrawal
-                block_wallet_deltas = db_wallet_delta[db_wallet_delta["block_number"] == block_number]
-                assert len(block_wallet_deltas) == expected_number_of_deltas
-                withdrawal_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == "WITHDRAWAL_SHARE"]
-                base_delta_df = block_wallet_deltas[block_wallet_deltas["base_token_type"] == BASE_TOKEN_SYMBOL]
-                assert len(withdrawal_delta_df) == 1
-                assert len(base_delta_df) == 1
-                withdrawal_delta = withdrawal_delta_df.iloc[0]
-                base_delta = base_delta_df.iloc[0]
-                # TODO check against expected withdrawal shares
-                assert withdrawal_delta["delta"] == -actual_num_withdrawal
-                # TODO check base delta
-                # TODO check wallet info matches the deltas
-                # TODO check pool info after this tx
-
-        # Check final wallet positions
-        db_current_wallet: pd.DataFrame = get_current_wallet(fast_hyperdrive_fixture.db_session, coerce_float=False)
-        # TODO currently only shorts are not dependent on poolinfo, so we only check shorts here
-        # Eventually we want to double check all token types
-        short_pos = db_current_wallet[db_current_wallet["base_token_type"] == "SHORT"]
-        assert short_pos.iloc[0]["value"] == Decimal(333)
-
-        # Check spot price and fixed rate
-        db_pool_analysis: pd.DataFrame = get_pool_analysis(fast_hyperdrive_fixture.db_session, coerce_float=False)
-        # Compare last value to what hyperdrive interface is reporting
-        interface = fast_hyperdrive_fixture.interface
-        latest_pool_analysis = db_pool_analysis.iloc[-1]
-
-        latest_spot_price = FixedPoint(str(latest_pool_analysis["spot_price"]))
-        expected_spot_price = interface.calc_spot_price()
-
-        latest_fixed_rate = FixedPoint(str(latest_pool_analysis["fixed_rate"]))
-        expected_fixed_rate = interface.calc_spot_rate()
-
-        assert latest_pool_analysis["block_number"] == interface.current_pool_state.block_number
-        # Spot price and fixed rate is off by two wei
-        assert isclose(latest_spot_price, expected_spot_price, abs_tol=FixedPoint("2e-18"))
-        assert isclose(latest_fixed_rate, expected_fixed_rate, abs_tol=FixedPoint("2e-18"))
